Subject: ANNOUNCE: Multilevel undo/redo for Gtk
I would like to announce the first public release of
my Gtk undo library (libundo version 0.1).

This library provides a Gtk object that manages multilevel
undo/redo for an application. The object stores a sequence
of actions and provides methods for undoing or redoing
those actions.  Actions can be grouped, so that multiple
primitive actions invoked in response to a single user
action can be undone/redone as a group. The object fires
signals to inform the application when it is or is not
possible to undo or redo any actions.  Utility functions
are provided that use those signals to keep user interface
components in sync with the undo sequence.

There is no documentation yet, but the header files are
heavily commented and the package includes unit tests and
an example Gtk application.

The library is available under the Gnu Lesser General
Public License.

It can be downloaded from:
     http://www-dse.doc.ic.ac.uk/~np2/software/undo.html

+------------------------------------------+---------------------+
| Name:   Nat Pryce MEng ACGI              | Dept. of Computing, |
| Email:  np2@doc.ic.ac.uk                 | Imperial College,   |
| Tel:    +44 (0)171 594 8394              | 180 Queen's Gate,   |
| Fax:    +44 (0)171 581 8024              | London SW7 2BZ,     |
| WWW:    http://www-dse.doc.ic.ac.uk/~np2 | United Kingdom      |
+------------------------------------------+---------------------+