Subject: ANNOUNCE: Gundo 0.3 undo/redo library for Gtk
Version 0.3 of Gundo, the undo/redo library, has been released
at http://www-dse.doc.ic.ac.uk/~np2/software/undo.html .

This new version includes API documentation generated from the
source code by Doc++.

 From the web page:
"Gundo is a Gtk object that manages multilevel undo/redo for an application.
The object stores a sequence of actions and provides methods for undoing or
redoing those actions. It fires signals to inform the application when it is
or is not possible to undo or redo any actions. Utility functions are
provided that react to those signals to keep user interface components in
sync with the undo sequence."

+------------------------------------------+---------------------+
| Name:   Nat Pryce MEng ACGI              | Dept. of Computing, |
| Email:  np2@doc.ic.ac.uk                 | Imperial College,   |
| Tel:    +44 (0)171 594 8394              | 180 Queen's Gate,   |
| Fax:    +44 (0)171 581 8024              | London SW7 2BZ,     |
| WWW:    http://www-dse.doc.ic.ac.uk/~np2 | United Kingdom      |
+------------------------------------------+---------------------+